fast-api/requirements.txt
streamlit
uvicorn
jwt
starlette
fastapi
psycopg2_binary
pydantic
pandas
numpy
pycaret